// packages/contexts/ingestion/src/blockchains/solana/types.ts
export interface SolscanTransaction {
  blockTime: number;
  fee: number;
  inputAccount: {
    account: string;
    postBalance: number;
    preBalance: number;
    signer: boolean;
    writable: boolean;
  }[];
  lamport: number;
  logMessage: string[];
  parsedInstruction: {
    params?: Record<string, unknown>;
    program: string;
    programId: string;
    type: string;
  }[];
  recentBlockhash: string;
  signer: string[];
  slot: number;
  status: 'Success' | 'Fail';
  txHash: string;
}